refactor: Extract the compiler_pipeline module into a separate file
